owasp-top10-2021-apps/a7/saidajaula-monster/app/requirements.txt
Click==8.1.3
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mysqlclient==2.0.3
six==1.11.0
visitor==0.1.3
Werkzeug==0.14.1
PyJWT